#1964 adding -Title parameter to Get-PnPAlert to retrieve alerts by title adding unit tests for Get-PnPAlert and Remove-PnPAlert moving unit tests to PrincipalsTests.cs since they don't belong to the WebTests where a sub web will be created each time (not needed) fixing minor spelling errors fixing ifdefs for unit tests to match those of the cmdlet files
